% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_timeseries.R
\name{cron_trend_close}
\alias{cron_trend_close}
\title{Call this function from the commandline (i.e. cron) to
store the current time-to-close metrics to the DB for
all known repos in the DB}
\usage{
cron_trend_close()
}
\description{
Call this function from the commandline (i.e. cron) to
store the current time-to-close metrics to the DB for
all known repos in the DB
}
